fix: binaryContent 삭제 시 존재 여부 확인 안되는 오류 수정

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -43,7 +43,9 @@ public class BasicBinaryContentService implements BinaryContentService {
 
     @Override
     public void delete(UUID binaryContentId) {
-        binaryContentRepository.existsById(binaryContentId);
+        if (!binaryContentRepository.existsById(binaryContentId)) {
+            throw new NoSuchElementException("[ERROR] 존재하지 않는 상태입니다.");
+        }
         binaryContentRepository.delete(binaryContentId);
     }
 }
